06-04-26/signup-login-jwt-oauth/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
streamlit
requests